Fixed bug with subdivide
SubdivisionFace::calculateFacePoint returned 0, this wasn't taken into
account in subdivide
common.pri now has LIBS variable as well as INCLUDEPATH
test case for SubdivisionFace::subdivide with 4 or more vertex points

+++ ShipCADlib/subdivcontrolcurve.cpp
@@ -288,7 +288,7 @@ void SubdivisionControlCurve::draw(Viewport &vp, LineShader* lineshader)
                     }
                     vertices << p3d2;
                 }
-                for (size_t j=0; j<curvelines.size(); ++j)
+                for (int j=0; j<curvelines.size(); ++j)
                     vertices << curvelines[j];
             }
             else {
